Update UCHARDET code-base

+++ uchardet/uchardet/src/JpCntx.cpp
@@ -1,4 +1,6 @@
-/* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*- */
+/* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*-
+ * vim: et sw=2 ts=2 fdm=marker
+ */
 /* ***** BEGIN LICENSE BLOCK *****
  * Version: MPL 1.1/GPL 2.0/LGPL 2.1
  *
@@ -133,14 +135,14 @@ void JapaneseContextAnalysis::HandleData(const char* aBuf, PRUint32 aLen)
   PRUint32 charLen;
   PRInt32 order;
   PRUint32 i;
-  
+
   if (mDone)
     return;
 
   //The buffer we got is byte oriented, and a character may span in more than one
-  //buffers. In case the last one or two byte in last buffer is not complete, we 
+  //buffers. In case the last one or two byte in last buffer is not complete, we
   //record how many byte needed to complete that character and skip these bytes here.
-  //We can choose to record those bytes as well and analyse the character once it 
+  //We can choose to record those bytes as well and analyse the character once it
   //is complete, but since a character will not make much difference, by simply skipping
   //this character will simply our logic and improve performance.
   for (i = mNeedToSkipCharNum; i < aLen; )
@@ -151,7 +153,7 @@ void JapaneseContextAnalysis::HandleData(const char* aBuf, PRUint32 aLen)
       mNeedToSkipCharNum = i - aLen;
       mLastCharOrder = -1;
     }
-    else 
+    else
     {
       if (order != -1 && mLastCharOrder != -1)
       {
@@ -161,12 +163,12 @@ void JapaneseContextAnalysis::HandleData(const char* aBuf, PRUint32 aLen)
           mDone = PR_TRUE;
           break;
         }
-        mRelSample[jp2CharContext[mLastCharOrder][order]]++;
+        mRelSample[(int)jp2CharContext[mLastCharOrder][order]]++;
       }
       mLastCharOrder = order;
     }
   }
-  
+
   return;
 }
 
@@ -187,7 +189,7 @@ float  JapaneseContextAnalysis::GetConfidence(void)
   //This is just one way to calculate confidence. It works well for me.
   if (mTotalRel > mDataThreshold)
     return ((float)(mTotalRel - mRelSample[0]))/mTotalRel;
-  else 
+  else
     return (float)DONT_KNOW;
 }
 
@@ -196,15 +198,15 @@ PRInt32 SJISContextAnalysis::GetOrder(const char* str, PRUint32 *charLen)
 {
   //find out current char's byte length
   if (((unsigned char)*str >= (unsigned char)0x81 && (unsigned char)*str <= (unsigned char)0x9f) ||
-      ((unsigned char)*str >= (unsigned char)0xe0 && (unsigned char)*str <= (unsigned char)0xfc))
+      ((unsigned char)*str >= (unsigned char)0xe0 && (unsigned char)*str <= (unsigned char)0xfc) )
       *charLen = 2;
   else
       *charLen = 1;
 
   //return its order if it is hiragana
   if (*str == '\202' &&
-      (unsigned char)*(str+1) >= (unsigned char)0x9f &&
-      (unsigned char)*(str+1) <= (unsigned char)0xf1)
+        (unsigned char)*(str+1) >= (unsigned char)0x9f &&
+        (unsigned char)*(str+1) <= (unsigned char)0xf1)
     return (unsigned char)*(str+1) - (unsigned char)0x9f;
   return -1;
 }
@@ -217,9 +219,9 @@ PRInt32 EUCJPContextAnalysis::GetOrder(const char* str, PRUint32 *charLen)
        (unsigned char)*str <= (unsigned char)0xfe))
       *charLen = 2;
   else if ((unsigned char)*str == (unsigned char)0x8f)
-      *charLen = 3;
+    *charLen = 3;
   else
-      *charLen = 1;
+    *charLen = 1;
 
   //return its order if it is hiragana
   if ((unsigned char)*str == (unsigned char)0xa4 &&
@@ -228,3 +230,5 @@ PRInt32 EUCJPContextAnalysis::GetOrder(const char* str, PRUint32 *charLen)
      return (unsigned char)*(str+1) - (unsigned char)0xa1;
   return -1;
 }
+
+
